+++ lib/src/search_app_bar_page/ui/infra/rx_get_type.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:get_state_manager/get_state_manager.dart';
+import 'package:get/state_manager.dart';
 
 typedef AuthFalseWidget = Widget Function();